requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.0
python-dotenv>=0.19.0
deep-translator>=1.11.0
gtts>=2.3.0
Pillow>=9.0.0
